Move Account.tsx to layout
https://github.com/litentry/litentry-app/pull/135#discussion_r655412617

fixes #127
